refactor: 기수 선택 리팩토링
